gen_video/api/requirements.txt
# API服务依赖
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# 任务队列
celery>=5.3.0
redis>=5.0.0

# 数据库
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9

# 配置管理
pyyaml>=6.0.1

# 监控
prometheus-client>=0.19.0

# 工具
httpx>=0.25.0
aiofiles>=23.2.0

